backend/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1.6
django-environ==0.11.2
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.4
